feat: :fire: Parallel grid search script
Created script to perform a grid search of 5 different classification algorithms and evaluate its performance. It uses a multiprocessing squema to run the grid search process with different dataset at the same time. This code outputs the evaluation metrics results and some plots to visualize and compare the perfomance of the algorithms.